+++ src/sysmon/constants.py
@@ -8,14 +8,14 @@ import platform
 import datetime
 
 # Version Information
-VERSION = "0.5.9b"
+VERSION = "0.5.9c"
 RELEASE_DATE = "2026-03-05"
-RELEASE_TIME = "0830 CST"
+RELEASE_TIME = "0957 CST"
 FULL_VERSION = f"v{VERSION} {RELEASE_DATE} {RELEASE_TIME}"
 
 # Build Information
 BUILD_DATE = "2026-03-05"
-BUILD_TIME = "0830 CST"
+BUILD_TIME = "0957 CST"
 BUILD_INFO = f"{BUILD_DATE} {BUILD_TIME}"
 
 # Runtime Information